% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_citations.R
\name{check_citations}
\alias{check_citations}
\title{Check all installed Bioconductor packages for missing CITATION files or DOIs.}
\usage{
check_citations(retrieve_missing_DOIs = TRUE)
}
\arguments{
\item{retrieve_missing_DOIs}{A logical value indicating whether to
retrieve missing DOIs from CrossRef.}
}
\value{
A data.frame of packages with CITATION files and DOIs.
}
\description{
Check all installed Bioconductor packages for missing CITATION files or DOIs.
}
\examples{
\dontrun{
check_citations()
}

}
