% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCitation.R
\name{updateCitation}
\alias{updateCitation}
\title{Create or update CITATION file using a DOI.}
\usage{
updateCitation(doi, output_file = "inst/CITATION")
}
\arguments{
\item{doi}{A character string representing the DOI of the article.}

\item{output_file}{A character string representing the path to the output file.
Default is "inst/CITATION".}
}
\value{
A CITATION file is created in the inst/ directory of the package.
}
\description{
This function creates a CITATION file from a DOI.
The CITATION file is created in the inst/ directory of the package.
}
\examples{
\dontrun{
updateCitation("10.1038/s41586-020-2818-3")
}

}
