% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_citation.R
\name{use_citation}
\alias{use_citation}
\title{Create or update CITATION file using a DOI.}
\usage{
use_citation(doi, output = "inst/CITATION")
}
\arguments{
\item{doi}{A character string representing the DOI of the article.}

\item{output}{A character string representing the path to the output file.
Default is "inst/CITATION".}
}
\value{
A CITATION file is created in the inst/ directory of the package.
}
\description{
This function creates a CITATION file from a DOI.
The CITATION file is created in the inst/ directory of the package.
}
\examples{
\dontrun{
use_citation("10.1038/s41586-020-2818-3")
}

}
